@@ -1,6 +1,6 @@
 import Vapor
 
-public protocol RefreshToken {
+public protocol RefreshToken: Sendable {
     var tokenString: String { get set }
     var clientID: String { get set }
     var userID: String? { get set }
