@@ -1,6 +1,6 @@
 import Vapor
 
-public protocol AuthorizeHandler {
+public protocol AuthorizeHandler: Sendable {
     func handleAuthorizationRequest(
         _ request: Request,
         authorizationRequestObject: AuthorizationRequestObject
@@ -8,14 +8,14 @@ public protocol AuthorizeHandler {
     func handleAuthorizationError(_ errorType: AuthorizationError) async throws -> Response
 }
 
-public enum AuthorizationError: Error {
+public enum AuthorizationError: Error, Sendable {
     case invalidClientID
     case confidentialClientTokenGrant
     case invalidRedirectURI
     case httpRedirectURI
 }
 
-public struct AuthorizationRequestObject {
+public struct AuthorizationRequestObject: Sendable {
     public let responseType: String
     public let clientID: String
     public let redirectURI: URI
