@@ -1,3 +1,3 @@
-public protocol ClientRetriever {
+public protocol ClientRetriever: Sendable {
     func getClient(clientID: String) async throws -> OAuthClient?
 }
