@@ -1,3 +1,3 @@
-public protocol ResourceServerRetriever {
+public protocol ResourceServerRetriever: Sendable {
     func getServer(_ username: String) async throws -> OAuthResourceServer?
 }
